requirements.txt
streamlit
rapidocr-onnxruntime
opencv-python-headless
pandas
pillow
numpy
opencc-python-reimplemented
